+++ Src/AssemblyInfo.cs
@@ -4,10 +4,10 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyProduct("SplitPipeline")]
-[assembly: AssemblyVersion("0.3.2")]
+[assembly: AssemblyVersion("0.4.0")]
 [assembly: AssemblyTitle("SplitPipeline - Parallel Data Processing in PowerShell")]
 [assembly: AssemblyCompany("https://github.com/nightroman/SplitPipeline")]
-[assembly: AssemblyCopyright("Copyright (c) 2011 Roman Kuzmin")]
+[assembly: AssemblyCopyright("Copyright (c) 2011-2012 Roman Kuzmin")]
 
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]